+++ src/app/backend/search/source-result.tsx
@@ -1,0 +1,8 @@
+import SearchResult from './search-result'
+
+interface SourceResult {
+    results: SearchResult[],
+    sourceName: string
+}
+
+export default SourceResult